Added new files

+++ largestthree.py
@@ -1,7 +1,7 @@
 # Finding the Largest of Three Numbers
-a = int(input("Enter a number:"))
-b = int(input("Enter a number:"))
-c = int(input("Enter a number:"))
+a = float(input("Enter a number:"))
+b = float(input("Enter a number:"))
+c = float(input("Enter a number:"))
 largest = a
 if (b > largest):
     largest = b
